Deynamic forms

+++ src/app/admin/dashboard/dashboard.component.ts
@@ -1,5 +1,6 @@
 import { Component } from '@angular/core';
 import { catchError } from 'rxjs';
+import { UserDataModelService } from 'src/app/auth/user-data-model.service';
 import { UserdataService } from 'src/app/shared/api/userdata.service';
 import { UserauthguardService } from 'src/app/shared/userauthguard.service';
 
@@ -20,7 +21,16 @@ export class DashboardComponent {
     this.userservice.getAllUserData()
     .pipe(
       catchError(error => {
-        return this.errorMessage = error.message;
+        let errorMsg = '';
+        if (error.error instanceof ErrorEvent) {
+            console.log('This is client side error');
+            errorMsg = `Error: ${error.error.message}`;
+        } else {
+            console.log('This is server side error');
+            errorMsg = `Error Code: ${error.status},  Message: ${error.message}`;
+        }
+        this.errorMessage = errorMsg
+        return this.errorMessage
       })
     )
     .subscribe( data => {
@@ -36,8 +46,9 @@ export class DashboardComponent {
   }
 
   delete(id:any){
-    this.userservice.deleteUserData(id)
-    window.location.reload()
+    if (confirm("Wanna delete this data?")){
+      this.userservice.deleteUserData(id)
+    }
   }
 
   update(data:any, id:any){
